Remove cyclic dep
